EV-INT-PR-SS/backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Async DB
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# MongoDB
motor==3.4.0

# Redis
redis[hiredis]==5.0.4

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Task queue
celery==5.4.0

# Object storage
boto3==1.34.100

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# HTTP client
httpx==0.27.0

# PDF
reportlab==4.1.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
hypothesis==6.100.1
httpx==0.27.0
factory-boy==3.3.0
aiosqlite==0.20.0
openpyxl>=3.1.0
